fix: division is float

+++ math_api/math_api/math/operations.py
@@ -12,7 +12,7 @@ def div(t1: int, t2: int) -> int:
     """For two integers t1 and t2, returns t1/t2."""
     if t2 == 0:
         raise ValueError("Division by 0")
-    return int(t1 / t2)
+    return t1 / t2
 
 
 def mul(t1: int, t2: int) -> int:
